--- src/main/java/kitchenpos/application/MenuService.java
@@ -1,52 +1,0 @@
-package kitchenpos.application;
-
-import java.util.List;
-import java.util.stream.Collectors;
-import kitchenpos.application.response.MenuResponse;
-import kitchenpos.domain.Menu;
-import kitchenpos.domain.repository.MenuGroupRepository;
-import kitchenpos.domain.repository.MenuRepository;
-import kitchenpos.domain.repository.ProductRepository;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-@Service
-public class MenuService {
-    private final MenuRepository menuRepository;
-    private final MenuGroupRepository menuGroupRepository;
-    private final ProductRepository productRepository;
-
-    public MenuService(MenuRepository menuRepository,
-                       MenuGroupRepository menuGroupRepository,
-                       ProductRepository productRepository) {
-        this.menuRepository = menuRepository;
-        this.menuGroupRepository = menuGroupRepository;
-        this.productRepository = productRepository;
-    }
-
-    @Transactional
-    public MenuResponse create(final Menu menu) {
-        if (!menuGroupRepository.existsById(menu.getMenuGroupId())) {
-            throw new IllegalArgumentException();
-        }
-
-        final List<Long> productIds = menu.getMenuProducts().stream()
-            .map(menuProduct -> menuProduct.getProduct().getId())
-            .collect(Collectors.toList());
-
-        final long count = productRepository.countProductInIds(productIds);
-
-        if(count != productIds.size()) {
-            throw new IllegalArgumentException();
-        }
-
-        return MenuResponse.from(menuRepository.save(menu));
-    }
-
-    public List<MenuResponse> list() {
-        return menuRepository.findAll()
-            .stream()
-            .map(MenuResponse::from)
-            .collect(Collectors.toList());
-    }
-}
