+++ src/main/java/kitchenpos/menu/application/MenuService.java
@@ -1,0 +1,35 @@
+package kitchenpos.menu.application;
+
+import java.util.List;
+import java.util.stream.Collectors;
+import kitchenpos.menu.application.response.MenuResponse;
+import kitchenpos.menu.application.validator.MenuValidator;
+import kitchenpos.menu.domain.Menu;
+import kitchenpos.menu.domain.repository.MenuRepository;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+public class MenuService {
+    private final MenuRepository menuRepository;
+    private final MenuValidator menuValidator;
+
+    public MenuService(MenuRepository menuRepository,
+                       MenuValidator menuValidator) {
+        this.menuRepository = menuRepository;
+        this.menuValidator = menuValidator;
+    }
+
+    @Transactional
+    public MenuResponse create(final Menu menu) {
+        menuValidator.validateCreation(menu);
+        return MenuResponse.from(menuRepository.save(menu));
+    }
+
+    public List<MenuResponse> list() {
+        return menuRepository.findAll()
+            .stream()
+            .map(MenuResponse::from)
+            .collect(Collectors.toList());
+    }
+}
