+++ src/main/java/kitchenpos/menu/application/response/MenuResponse.java
@@ -1,9 +1,9 @@
-package kitchenpos.application.response;
+package kitchenpos.menu.application.response;
 
 import java.math.BigDecimal;
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.domain.Menu;
+import kitchenpos.menu.domain.Menu;
 
 public class MenuResponse {
