+++ src/main/java/kitchenpos/menu/domain/repository/MenuRepository.java
@@ -1,7 +1,7 @@
-package kitchenpos.domain.repository;
+package kitchenpos.menu.domain.repository;
 
 import java.util.List;
-import kitchenpos.domain.Menu;
+import kitchenpos.menu.domain.Menu;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
