+++ src/main/java/kitchenpos/product/application/ProductService.java
@@ -27,4 +27,8 @@ public class ProductService {
             .map(ProductResponse::from)
             .collect(Collectors.toList());
     }
+
+    public long countProductInIds(List<Long> productIds) {
+        return productRepository.countProductInIds(productIds);
+    }
 }
