+++ src/main/java/kitchenpos/ui/MenuGroupRestController.java
@@ -2,10 +2,9 @@ package kitchenpos.ui;
 
 import java.net.URI;
 import java.util.List;
-import kitchenpos.application.MenuGroupService;
-import kitchenpos.domain.MenuGroup;
+import kitchenpos.menu.application.MenuGroupService;
 import kitchenpos.ui.request.MenuGroupCreateRequest;
-import kitchenpos.application.response.MenuGroupResponse;
+import kitchenpos.menu.application.response.MenuGroupResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
