+++ src/main/java/kitchenpos/ui/MenuRestController.java
@@ -1,10 +1,8 @@
 package kitchenpos.ui;
 
-import java.util.stream.Collectors;
-import kitchenpos.application.MenuService;
-import kitchenpos.domain.Menu;
+import kitchenpos.menu.application.MenuService;
 import kitchenpos.ui.request.MenuCreateRequest;
-import kitchenpos.application.response.MenuResponse;
+import kitchenpos.menu.application.response.MenuResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
