+++ src/main/java/kitchenpos/ui/request/MenuCreateRequest.java
@@ -3,9 +3,9 @@ package kitchenpos.ui.request;
 import java.math.BigDecimal;
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.domain.Menu;
-import kitchenpos.domain.MenuGroup;
-import kitchenpos.domain.MenuProduct;
+import kitchenpos.menu.domain.Menu;
+import kitchenpos.menu.domain.MenuGroup;
+import kitchenpos.menu.domain.MenuProduct;
 
 public class MenuCreateRequest {
