+++ src/main/java/kitchenpos/ui/request/MenuGroupCreateRequest.java
@@ -1,6 +1,6 @@
 package kitchenpos.ui.request;
 
-import kitchenpos.domain.MenuGroup;
+import kitchenpos.menu.domain.MenuGroup;
 
 public class MenuGroupCreateRequest {
