+++ src/main/java/kitchenpos/ui/request/OrderLineItemRequest.java
@@ -1,6 +1,6 @@
 package kitchenpos.ui.request;
 
-import kitchenpos.domain.Menu;
+import kitchenpos.menu.domain.Menu;
 import kitchenpos.domain.OrderLineItem;
 
 public class OrderLineItemRequest {
