+++ src/test/java/kitchenpos/integration/api/MenuGroupApi.java
@@ -4,7 +4,7 @@ import java.util.List;
 import kitchenpos.integration.utils.MockMvcResponse;
 import kitchenpos.integration.utils.MockMvcUtils;
 import kitchenpos.ui.request.MenuGroupCreateRequest;
-import kitchenpos.application.response.MenuGroupResponse;
+import kitchenpos.menu.application.response.MenuGroupResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
